QUEUE IS WORKING PROPERLY

+++ ajax/processqueue.php
@@ -15,17 +15,28 @@
 				$row = pg_fetch_row($result);
 			if($result)					
 			{
-
-
 				$user = $_SESSION['logged_admin'];
+				if(pg_num_rows($result)>0)
+				{
 
-				$updQuery = "UPDATE admin SET is_processing=$row[0] where username ='$user'";
-				$result = pg_query($dbconn,$updQuery);
-				$delQuery = "DELETE from QUEUE where (slot_id=$row[0] and student_id=$row[1] and day_id=$row[2])";
-				$result = pg_query($dbconn,$delQuery);
+					$updQuery = "UPDATE admin SET is_processing=$row[0] where username ='$user'";
+					$result = pg_query($dbconn,$updQuery);
+					$delQuery = "DELETE from QUEUE where (slot_id=$row[0] and student_id=$row[1] and day_id=$row[2])";
+					$resultDel = pg_query($dbconn,$delQuery);
+				}
+				else
+				{
+
+					$updQuery2 = "UPDATE admin SET is_processing=NULL where username ='$user'";
+					$resultUpd = pg_query($dbconn,$updQuery2);
+				}
 			}
 			else
 			{
+
+					$updQuery2 = "UPDATE admin SET is_processing=NULL where username ='$user'";
+					$resultUpd = pg_query($dbconn,$updQuery2);
+
 				echo "NO MORE STUDENTS TO PROCESS!";
 			}
 	}
